# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management Platform API
  version: 0.2.0
  description: |
    Refactor contract-first API baseline.
    Due date semantic migration: `dueAt` is canonical; `dueDate` accepted for backward compatibility in write payloads and exposed only in legacy `/tasks`.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Paged project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}:
    get:
      summary: Get project detail
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      summary: Delete project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Deleted project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/timeline:
    get:
      summary: Get project timeline projection
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project timeline items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTimelineResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /projects/{projectId}/archive:
    post:
      summary: Archive a project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Archived project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/member-candidates:
    get:
      summary: Search users who can be added to a project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - in: query
          name: q
          schema: { type: string }
          description: Optional name or email search text
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Users who are not already project members
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemberCandidate'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/members:
    post:
      summary: Add or update a project member
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId: { type: string }
                role: { type: string, enum: [viewer, member, project_admin] }
      responses:
        '201':
          description: Project member
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ProjectMember'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{projectId}/milestones:
    post:
      summary: Create a project milestone
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                dueAt: { type: [string, 'null'], format: date-time }
      responses:
        '201':
          description: Created milestone
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Milestone'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{projectId}/sprints:
    post:
      summary: Create a project sprint
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                goal: { type: string }
                startAt: { type: string, format: date-time }
                endAt: { type: [string, 'null'], format: date-time }
      responses:
        '201':
          description: Created sprint
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Sprint'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Missing/invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthErrorResponse' }
    Forbidden:
      description: Role/scope forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Unique constraint conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ProjectMember:
      type: object
      required: [projectId, userId, role, name, email]
      properties:
        projectId: { type: string }
        userId: { type: string }
        role: { type: string, enum: [viewer, member, project_admin] }
        name: { type: string }
        email: { type: string, format: email }
    Milestone:
      type: object
      required: [id, projectId, name, status]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        dueAt: { type: [string, 'null'], format: date-time }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    Sprint:
      type: object
      required: [id, projectId, name, status]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        goal: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: [string, 'null'], format: date-time }
        status: { type: string }
        createdAt: { type: string, format: date-time }
...
    AuthErrorResponse:
      type: object
      properties:
        error:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            status: { type: integer }
    Project:
      type: object
      required: [id, key, name, description, ownerId, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        key: { type: string }
        name: { type: string }
        description: { type: string }
        ownerId: { type: [string, 'null'] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        members:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
        milestones:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              dueAt: { type: [string, 'null'], format: date-time }
        sprints:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
    MemberCandidate:
      type: object
      required: [id, name, email]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
    ProjectResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Project'
    ProjectListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        meta:
          type: object
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
    ProjectTimelineItem:
      type: object
      properties:
        type: { type: string, enum: [milestone, sprint, issue] }
        id: { type: string }
        name: { type: string }
        startAt: { type: [string, 'null'], format: date-time }
        endAt: { type: [string, 'null'], format: date-time }
        status: { type: [string, 'null'] }
    ProjectTimeline:
      type: object
      properties:
        project:
          type: object
          properties:
            id: { type: string }
            key: { type: string }
            name: { type: string }
            status: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectTimelineItem'
    ProjectTimelineResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ProjectTimeline'
        meta:
          type: object
          properties:
            lastSync: { type: string, format: date-time }
    CreateProjectRequest:
      type: object
      required: [key, name]
      properties:
        key: { type: string, pattern: '^[A-Z][A-Z0-9_-]{1,11}$' }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string }
    UpdateProjectRequest:
      type: object
      properties:
        key: { type: string, pattern: '^[A-Z][A-Z0-9_-]{1,11}$' }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string }
        status: { type: string }
...